src/rag/requirement.txt
semantic_chunkers
weaviate-client
underthesea